--- hubspot/close.go
@@ -1,5 +1,0 @@
-package hubspot
-
-func (c *Connector) Close() error {
-	return nil
-}
