--- hubspot/name.go
@@ -1,6 +1,0 @@
-package hubspot
-
-// Name returns the name of the connector.
-func (c *Connector) Name() string {
-	return c.Provider().String()
-}
