+++ salesforce/client.go
@@ -10,3 +10,7 @@ func (c *Connector) JSONHTTPClient() *common.JSONHTTPClient {
 func (c *Connector) HTTPClient() *common.HTTPClient {
 	return c.Client.HTTPClient
 }
+
+func (c *Connector) Close() error {
+	return nil
+}
