--- salesforce/close.go
@@ -1,5 +1,0 @@
-package salesforce
-
-func (c *Connector) Close() error {
-	return nil
-}
